signifant changes to working
Database insertion is working .
Small UI changes .
Many files added
