+++ Icons/icon_res.qrc
@@ -1,5 +1,6 @@
 <RCC>
   <qresource prefix="icons">
+    <file>close line.png</file>
     <file>blue.png</file>
     <file>blueish_grad.png</file>
     <file>graywood.png</file>
