--- README.md
@@ -1,1 +1,0 @@
-# Project_NCC
